feat: rating
* Add constraint for max no of attacks on a base per level

* Log attacker path

* Drop table attacker_path

* fix: Add new map layout for level if it does not exist

* fix: get user from session in replay route

+++ src/api/user/util.rs
@@ -13,8 +13,8 @@ use serde::Serialize;
 pub struct StatsResponse {
     pub highest_attack_score: i32,
     pub highest_defense_score: i32,
-    pub rating: i32,
-    pub highest_rating: i32,
+    pub rating: f32,
+    pub highest_rating: f32,
     pub position_in_leaderboard: i32,
     pub no_of_robots_killed: i32,
     pub no_of_robots_got_killed: i32,
